pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vaultforge"
version = "0.1.0"
description = "Local LLM-powered second brain / knowledge compiler"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "python-multipart>=0.0.9",
    "jinja2>=3.1",
    "markdown>=3.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
secondbrain = "secondbrain.cli:app"

[tool.setuptools.packages.find]
include = ["secondbrain*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
